Napari viewer: Add deferring adapter between ilastik GUI and napari layer
The "better" way to achieve lazy-loading images into Napari was to just pass the image layer the lazyflow output slot.
Due to slots' accessibility by slicing, this can work with a light Slot wrapper class that provides other parts of the numpy ndarray-API that Napari uses. But when sliced, slots return lazyflow Requests, which Napari would need to know when to `wait()` for.