wip: sussing out what causes boundary artifacts.
conclusion: artifacts around block boundaries only seem to be avoided when the halo is a multiple of the scaling factor.

I'm guessing that only then is the halo removal actually clean after scaling the raw block with halo down to the target size; otherwise the halo removed after downscaling is affected by rounding errors.

Try taking apart skimage.resize implementation. Rather than passing it the raw block with halo to downscale, and then removing a scaled halo -- apply the gaussian filter on the raw block with halo, then remove the halo, and then pass the blurred raw block to zoom